Adicao do Servico de EmailLog e troca do uso de Queues para Exchange

+++ src/GeekShooping.CouponAPI/Model/Coupon.cs
@@ -11,6 +11,6 @@ namespace GeekShopping.CouponAPI.Model
         [StringLength(30)]
         public string CouponCode { get; set; }
         [Required]        
-        public double DiscountAmount { get; set; }        
+        public decimal DiscountAmount { get; set; }        
     }
 }
