+++ src/GeekShopping.CartAPI/Messages/CheckoutHeaderDto.cs
@@ -7,8 +7,8 @@ namespace GeekShopping.CartAPI.Messages
     {
         public string UserId { get; set; }
         public string CouponCode { get; set; }
-        public double PurchaseAmount { get; set; }
-        public double DiscountAmount { get; set; }
+        public decimal PurchaseAmount { get; set; }
+        public decimal DiscountAmount { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public DateTime Time { get; set; }
