+++ src/GeekShopping.Web/Models/CartHeaderViewmModel.cs
@@ -5,8 +5,8 @@
         public long Id { get; set; }
         public string UserId { get; set; }
         public string CouponCode { get; set; }
-        public double PurchaseAmount { get; set; }
-        public double DiscountAmount { get; set; }
+        public decimal PurchaseAmount { get; set; }
+        public decimal DiscountAmount { get; set; }
         public string FirstName{ get; set; }
         public string LastName{ get; set; }
         public DateTime Time{ get; set; }
